requirements.txt
altgraph==0.17.4
packaging==25.0
pefile==2023.2.7
pycryptodome==3.23.0
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.6
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pywin32-ctypes==0.2.3
setuptools==80.9.0